+++ TeaNotes/Notes/Models/TeaNoteImpression.cs
@@ -4,6 +4,6 @@
     {
         public required string Comment { get; set; }
         public required string WellCombinedWith { get; set; }
-        public required double? Rate { get; set; }
+        public required int? Rate { get; set; }
     }
 }
